Fix bayesian generator model creation, move io.py to base.py
