+++ tests/test_io.py
@@ -3,9 +3,8 @@ from copy import copy
 import yaml
 
 from xopt.evaluator import Evaluator
-from xopt.base import Xopt
+from xopt.base import Xopt, parse_config, state_to_dict
 from xopt.generators import RandomGenerator
-from xopt.io import parse_config, state_to_dict
 from xopt.resources.testing import TEST_VOCS_BASE, TEST_YAML
